Split out methods to facilitate testing
